Use the archive with precompiled binaries for updating the resource dir
While most headers in the resource directory are present in the clang
source archive, some are generated during compiling.

As a bonus, non-header files do not exist in the resource directory, in
the archives containing binaries. Previously we have manually removed a
CMakeLists.txt, but have been leaving behind some READMEs.

HLSL headers are not installed, because the feature is highly
experimental.
openmp_wrappers/{stdlib.h,time.h} are not intended to be included in resource directory.